requirements.txt
pandas
pydantic
python-dotenv
streamlit
psycopg2-binary
